requirements.txt
# Core Scraping
scrapy==2.11.0
scrapy-playwright==0.0.34
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Orchestration
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Utilities
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
dateparser==1.2.0
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Data Export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Anti-bot measures
fake-useragent==1.4.0
randomheaders==1.1.2

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Validation
validators==0.22.0
cerberus==1.3.5

# CLI
click==8.1.7
rich==13.7.0
